+++ tests/Unit/Drivers/GhasedakDriverTest.php
@@ -39,6 +39,7 @@ final class GhasedakDriverTest extends TestCase
         Http::assertSent(fn (Request $request) => $request->hasHeader('ApiKey', 'sms_token')
             && $request->hasHeader('Content-Type', 'application/json')
             && $request->url() === 'https://gateway.ghasedak.me/rest/api/v1/WebService/end-point'
+            && $request->method() === 'POST'
             && $request['key'] === 'value');
     }